separates common HITRAN/HITEMP
