+++ tests/unittests/spec/api/api_test.py
@@ -1,8 +1,8 @@
-from exojax.database.api  import _convert_proper_isotope
-from exojax.database.api  import _isotope_index_from_isotope_number
-from exojax.database.api  import _QT_interp
-from exojax.database.api  import _qr_interp
-from exojax.database.api  import _qr_interp_lines
+from exojax.database._common.isotope_functions import _convert_proper_isotope
+from exojax.database._common.isotope_functions import _isotope_index_from_isotope_number
+from exojax.database._common.partition_function import _QT_interp
+from exojax.database._common.partition_function import _qr_interp
+from exojax.database._common.partition_function import _qr_interp_lines
 from exojax.test.emulate_mdb import mock_mdbHitemp
 from exojax.utils.constants import Tref_original
 import numpy as np
@@ -41,8 +41,8 @@ def test__qr_interp_lines():
     mdb = mock_mdbHitemp(multi_isotope=True)
     T = 1000.0
     val = np.mean(
-        _qr_interp_lines(T, mdb.isoid, mdb.uniqiso, mdb.T_gQT, mdb.gQT,
-                         Tref_original))
+        _qr_interp_lines(T, mdb.isoid, mdb.uniqiso, mdb.T_gQT, mdb.gQT, Tref_original)
+    )
     print(val)
     assert val == pytest.approx(3.553517)
 
@@ -55,13 +55,13 @@ def test__exact_isotope_name():
 def test_molmass():
     mdb = mock_mdbHitemp(multi_isotope=True)
     assert mdb.molmass == pytest.approx(28.01044518292034)
-    #assert mdb.exact_isotope_name(1) == "(12C)(16O)"
+    # assert mdb.exact_isotope_name(1) == "(12C)(16O)"
 
 
 if __name__ == "__main__":
-        test__convert_proper_isotope()
-        test__isotope_index_from_isotope_number()
-        test__QT_interp()
-        test__qr_interp()
-        test__qr_interp_lines()
-        test_molmass()
+    test__convert_proper_isotope()
+    test__isotope_index_from_isotope_number()
+    test__QT_interp()
+    test__qr_interp()
+    test__qr_interp_lines()
+    test_molmass()
